refactor: component folder structure
